office_at_night.jpg.txt: Adjust to the values generated on the testbed
Despite the same yolov3-tiny.{cfg,weights} used locally the value seems to
differs a bit.  The opencv version used here is opencv-3.4.9nb4 (so a different
one), py38-numpy-1.18.5 and on NetBSD/amd64 9.99.64, of course.

The values seems to not differs significantly so better to avoid possibly ending
up in a whiterabbit hole to try to figure out what's really going on, sorry!